moved the doc_generator code into masters
