+++ apps/masters/urls.py
@@ -39,4 +39,6 @@ urlpatterns = [
     path('', include(router.urls)),
     path('uploads/', FileUploadView.as_view(), name='file_uploads'),
     path('generate_order_no/', generate_order_number_view, name='generate_order_no'),
+    path('document_generator/<str:pk>/<str:document_type>/', DocumentGeneratorView.as_view(), name='generate-sale-order-pdf'),
+
 ]
