+++ apps/masters/utils/docs_variables.py
